% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_cal.R
\name{neighbor_cal}
\alias{neighbor_cal}
\title{OASISAD calculates neighbor function}
\usage{
neighbor_cal(wm_pve, nei_seg, nei_wm, label, re_value)
}
\arguments{
\item{wm_pve}{A voxel's probability of being white matter}

\item{nei_seg}{Segmentation information of its neighbors}

\item{nei_wm}{White matter probability information of its neighbors}

\item{label}{White matter's label in segmentation mask}

\item{re_value}{User defined value to refine WMH probability}
}
\value{
Refined probability map
}
\description{
Input voxel's neigbhors' information to calculate refinement parameter
}
