% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_indx.R
\name{neighbor_indx}
\alias{neighbor_indx}
\title{OASISAD calculates neighbor indices function}
\usage{
neighbor_indx(x, dim)
}
\arguments{
\item{x}{Voxel's 3D indices}

\item{dim}{Dimension of 3D image}
}
\value{
Refined probability map
}
\description{
Input voxel's index information to calculate its neighbor indices
}
