% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisad_refine.R
\name{oasis_refine}
\alias{oasis_refine}
\title{OASISAD refine probability function}
\usage{
oasis_refine(
  prob_map,
  neighbor = FALSE,
  seg = NULL,
  wm = NULL,
  wm_label = NULL,
  re_value = NULL,
  indx = NULL,
  indx3d = NULL
)
}
\arguments{
\item{prob_map}{A dataframe includes a voxel's probability of being white matter
and its nearest neighor voxels' vetorized indice.}

\item{neighbor}{A boolean whether using neighor function to correct probability map}

\item{seg}{File path of segmentation mask segmentation tool such as FSL
Please set non-brain tissue in seg to 0}

\item{wm}{File path of white matter pve from segmentation tool such as FSL}

\item{wm_label}{A value indicates white matter label in input seg_voxel argument.
If null, it is 3, by FSL.}

\item{re_value}{A value used in neighbor calculation function to
suppress probability of being WMH}

\item{indx}{Voxels' vectorized indices}

\item{indx3d}{Voxels' 3D indices}
}
\value{
Refined probability map
}
\description{
Refine voxel's probability of being
white matter hypeintensity by neighbor information. Neighbor
}
