% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisad_df.R
\name{oasisad_df}
\alias{oasisad_df}
\title{OASIS Data Frame}
\usage{
oasisad_df(
  flair,
  t1,
  t2 = NULL,
  pd = NULL,
  gold_standard = NULL,
  preproc = TRUE,
  brain_mask = NULL,
  img_space = FALSE,
  neighbor = TRUE,
  wm_mask = NULL,
  seg_mask = NULL,
  dir = NULL,
  eroder = TRUE,
  voxel_select = NULL,
  normalize = TRUE,
  image_sm = TRUE,
  slices = NULL,
  orientation = c("axial", "coronal", "sagittal"),
  return_pre = FALSE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{flair}{Input FLAIR image}

\item{t1}{Input T1 image}

\item{t2}{Input T2 image}

\item{pd}{Input PD image}

\item{gold_standard}{gold standard lesion segmentation mask of class
\code{nifti}}

\item{preproc}{A boolean indicates whether to call
\code{oasisad_pre} function or not}

\item{brain_mask}{Input brain_mask, if null, a mask will be obtained by FSL}

\item{img_space}{An image to register, if NULL, 'flair' image will be used in registration.}

\item{neighbor}{A boolean indicates whether will use neighbor refinement function in model step.
If true, either prepare segmentation and white matter mask to input in this function or
this functoin will generate masks by \code{fslr}}

\item{wm_mask}{Input of white matter mask}

\item{seg_mask}{Input of segmentation mask}

\item{dir}{A user defined output path for fslr segmentation}

\item{eroder}{A boolean indicates whether should use \code{fslerode}}

\item{voxel_select}{A specifed level to remove voxels whose intensity under}

\item{normalize}{A boolean indicates whether to
perform z-score normalization of the image over the brain mask,
should be \code{TRUE} unless you train model
using an alternative normalization. Normoalization is a recommended step in image preprocessing.}

\item{image_sm}{A boolean indates whether to smooth images and used as predictors in model}

\item{slices}{vector of desired slices to train on, if \code{NULL}
then train over the entire brain mask}

\item{orientation}{string value telling which orientation the
training slices are specified in, can take the values of "axial",
"sagittal", or "coronal"}

\item{return_pre}{is a logical value that indicates whether
the preprocessed images should be returned}

\item{cores}{numeric indicating the number of cores to be used}

\item{verbose}{A boolean indicated whether output messages}
}
\value{
OASISAS data structure
}
\description{
This function creates data frame for OASIS models
}
