% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisad_eva.R
\name{oasisad_eva}
\alias{oasisad_eva}
\title{OASISAD evaluation function}
\usage{
oasisad_eva(pred, truth, pos_l = NULL)
}
\arguments{
\item{pred}{a vector includes predicted labels from model.}

\item{truth}{a vector includes true labels. Please check labels' names in pred and truth are same.}

\item{pos_l}{a value indicates the positive label. If \emph{NULL}, default will set last value from \emph{table} function as positive label.}
}
\value{
a table includes evaluation metrics.
}
\description{
Evaluate performance of segmentation models based on their outputs
}
