% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisad_opt_th.R
\name{opt_thre}
\alias{opt_thre}
\title{OASISAD optimal threshold selection function}
\usage{
opt_thre(probs, gs, cutoff)
}
\arguments{
\item{probs}{Probability map}

\item{gs}{Gold standard of input voxels}

\item{cutoff}{A numeric vector includes threshold candidates}
}
\value{
Refined probability map
}
\description{
This function will use an input cutoff list to select the optimal
cutoff with metric DSC(F1-score) based on validation data
}
