% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapply_pb.R
\name{sapply_pb}
\alias{sapply_pb}
\title{OASISAD sapply with processing bar}
\usage{
sapply_pb(X, FUN, ...)
}
\arguments{
\item{X}{same argument in sapply.}

\item{FUN}{same argument in sapply.}

\item{...}{Arguments for FUN}
}
\value{
results same as sapply
}
\description{
Add process bar to sapply.
}
