% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisad_pre.R
\name{oasisad_pre}
\alias{oasisad_pre}
\title{OASISAD image preprocessing function}
\usage{
oasisad_pre(
  flair,
  t1,
  t2 = NULL,
  pd = NULL,
  img_space = NULL,
  brain_mask = NULL,
  segmentation = FALSE,
  dir = NULL,
  cores = 1,
  verbose = NULL
)
}
\arguments{
\item{flair}{Input FLAIR image}

\item{t1}{Input T1 image}

\item{t2}{Input T2 image}

\item{pd}{Input PD image}

\item{img_space}{An image to register and use for creating brain mask if needed.
If NULL, 'flair' image will be used in registration.}

\item{brain_mask}{Input brain_mask, if null, a mask will be obtained by FSL}

\item{segmentation}{A boolean indicates whether using bias_correct for segmentation}

\item{dir}{A user defined output path for fslr segmentation}

\item{cores}{A number indicates how many cores used mclapply}

\item{verbose}{A boolean indicated whether output messages}
}
\value{
preprocessed image data
}
\description{
MRI image data preprocessing with multiple inputs
}
